ddl: move placement rule tests into another file
