+++ ddl/db_test.go
@@ -5096,114 +5096,6 @@ func (s *testSerialDBSuite) TestCommitTxnWithIndexChange(c *C) {
 	tk.MustExec("admin check table t1")
 }
 
-func (s *testDBSuite1) TestAlterTableAlterPartition(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t1")
-	defer tk.MustExec("drop table if exists t1")
-
-	tk.MustExec(`create table t1 (c int)
-PARTITION BY RANGE (c) (
-	PARTITION p0 VALUES LESS THAN (6),
-	PARTITION p1 VALUES LESS THAN (11),
-	PARTITION p2 VALUES LESS THAN (16),
-	PARTITION p3 VALUES LESS THAN (21)
-);`)
-
-	_, err := tk.Exec(`alter table t1 alter partition p0
-add placement policy
-	role=leader
-	replicas=3`)
-	c.Assert(err, ErrorMatches, ".*pd unavailable.*")
-
-	_, err = tk.Exec(`alter table t1 alter partition p0
-add placement policy
-	constraints='+zone=sh'
-	role=leader
-	replicas=3`)
-	c.Assert(err, ErrorMatches, ".*pd unavailable.*")
-
-	_, err = tk.Exec(`alter table t1 alter partition p0
-add placement policy
-	constraints='+   zone   =   sh  ,     - zone = bj    '
-	role=leader
-	replicas=3`)
-	c.Assert(err, ErrorMatches, ".*pd unavailable.*")
-
-	_, err = tk.Exec(`alter table t1 alter partition p0
-add placement policy
-	constraints=',,,'
-	role=leader
-	replicas=3`)
-	c.Assert(err, ErrorMatches, ".*constraint too short to be valid.*")
-
-	_, err = tk.Exec(`alter table t1 alter partition p0
-add placement policy
-	constraints='0000'
-	role=leader
-	replicas=3`)
-	c.Assert(err, ErrorMatches, ".*unknown operation.*")
-
-	_, err = tk.Exec(`alter table t1 alter partition p0
-add placement policy
-	constraints='+000'
-	role=leader
-	replicas=3`)
-	c.Assert(err, ErrorMatches, ".*invalid constraint format.*")
-
-	_, err = tk.Exec(`alter table t1 alter partition p0
-add placement policy
-	constraints='+    '
-	role=leader
-	replicas=3`)
-	c.Assert(err, ErrorMatches, ".*invalid constraint format.*")
-
-	_, err = tk.Exec(`alter table t1 alter partition p0
-add placement policy
-	constraints='+=zone1'
-	role=leader
-	replicas=3`)
-	c.Assert(err, ErrorMatches, ".*invalid constraint format.*")
-
-	_, err = tk.Exec(`alter table t1 alter partition p0
-add placement policy
-	constraints='+=    z'
-	role=leader
-	replicas=3`)
-	c.Assert(err, ErrorMatches, ".*invalid constraint format.*")
-
-	_, err = tk.Exec(`alter table t1 alter partition p0
-add placement policy
-	constraints='+zone='
-	role=leader
-	replicas=3`)
-	c.Assert(err, ErrorMatches, ".*invalid constraint format.*")
-
-	_, err = tk.Exec(`alter table t1 alter partition p0
-add placement policy
-	constraints='+z  ='
-	role=leader
-	replicas=3`)
-	c.Assert(err, ErrorMatches, ".*invalid constraint format.*")
-
-	_, err = tk.Exec(`alter table t1 alter partition p
-add placement policy
-	constraints='+zone=sh'
-	role=leader
-	replicas=3`)
-	c.Assert(err, ErrorMatches, ".*Unknown partition.*")
-
-	tk.MustExec("drop table if exists t1")
-	tk.MustExec("create table t1 (c int)")
-
-	_, err = tk.Exec(`alter table t1 alter partition p
-add placement policy
-	constraints='+zone=sh'
-	role=leader
-	replicas=3`)
-	c.Assert(err, ErrorMatches, ddl.ErrPartitionMgmtOnNonpartitioned)
-}
-
 func init() {
 	// Make sure it will only be executed once.
 	domain.SchemaOutOfDateRetryInterval = int64(50 * time.Millisecond)
