+++ ddl/placement_rule_test.go
@@ -1,0 +1,128 @@
+// Copyright 2015 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package ddl_test
+
+import (
+	. "github.com/pingcap/check"
+	"github.com/pingcap/tidb/ddl"
+	"github.com/pingcap/tidb/util/testkit"
+)
+
+func (s *testDBSuite1) TestAlterTableAlterPartition(c *C) {
+	tk := testkit.NewTestKit(c, s.store)
+	tk.MustExec("use test")
+	tk.MustExec("drop table if exists t1")
+	defer tk.MustExec("drop table if exists t1")
+
+	tk.MustExec(`create table t1 (c int)
+PARTITION BY RANGE (c) (
+	PARTITION p0 VALUES LESS THAN (6),
+	PARTITION p1 VALUES LESS THAN (11),
+	PARTITION p2 VALUES LESS THAN (16),
+	PARTITION p3 VALUES LESS THAN (21)
+);`)
+
+	_, err := tk.Exec(`alter table t1 alter partition p0
+add placement policy
+	role=leader
+	replicas=3`)
+	c.Assert(err, ErrorMatches, ".*pd unavailable.*")
+
+	_, err = tk.Exec(`alter table t1 alter partition p0
+add placement policy
+	constraints='+zone=sh'
+	role=leader
+	replicas=3`)
+	c.Assert(err, ErrorMatches, ".*pd unavailable.*")
+
+	_, err = tk.Exec(`alter table t1 alter partition p0
+add placement policy
+	constraints='+   zone   =   sh  ,     - zone = bj    '
+	role=leader
+	replicas=3`)
+	c.Assert(err, ErrorMatches, ".*pd unavailable.*")
+
+	_, err = tk.Exec(`alter table t1 alter partition p0
+add placement policy
+	constraints=',,,'
+	role=leader
+	replicas=3`)
+	c.Assert(err, ErrorMatches, ".*constraint too short to be valid.*")
+
+	_, err = tk.Exec(`alter table t1 alter partition p0
+add placement policy
+	constraints='0000'
+	role=leader
+	replicas=3`)
+	c.Assert(err, ErrorMatches, ".*unknown operation.*")
+
+	_, err = tk.Exec(`alter table t1 alter partition p0
+add placement policy
+	constraints='+000'
+	role=leader
+	replicas=3`)
+	c.Assert(err, ErrorMatches, ".*invalid constraint format.*")
+
+	_, err = tk.Exec(`alter table t1 alter partition p0
+add placement policy
+	constraints='+    '
+	role=leader
+	replicas=3`)
+	c.Assert(err, ErrorMatches, ".*invalid constraint format.*")
+
+	_, err = tk.Exec(`alter table t1 alter partition p0
+add placement policy
+	constraints='+=zone1'
+	role=leader
+	replicas=3`)
+	c.Assert(err, ErrorMatches, ".*invalid constraint format.*")
+
+	_, err = tk.Exec(`alter table t1 alter partition p0
+add placement policy
+	constraints='+=    z'
+	role=leader
+	replicas=3`)
+	c.Assert(err, ErrorMatches, ".*invalid constraint format.*")
+
+	_, err = tk.Exec(`alter table t1 alter partition p0
+add placement policy
+	constraints='+zone='
+	role=leader
+	replicas=3`)
+	c.Assert(err, ErrorMatches, ".*invalid constraint format.*")
+
+	_, err = tk.Exec(`alter table t1 alter partition p0
+add placement policy
+	constraints='+z  ='
+	role=leader
+	replicas=3`)
+	c.Assert(err, ErrorMatches, ".*invalid constraint format.*")
+
+	_, err = tk.Exec(`alter table t1 alter partition p
+add placement policy
+	constraints='+zone=sh'
+	role=leader
+	replicas=3`)
+	c.Assert(err, ErrorMatches, ".*Unknown partition.*")
+
+	tk.MustExec("drop table if exists t1")
+	tk.MustExec("create table t1 (c int)")
+
+	_, err = tk.Exec(`alter table t1 alter partition p
+add placement policy
+	constraints='+zone=sh'
+	role=leader
+	replicas=3`)
+	c.Assert(err, ErrorMatches, ddl.ErrPartitionMgmtOnNonpartitioned)
+}
